feat: integration with reward_hub
